MateMaTeX/requirements.txt
# Core framework
crewai[tools,google-genai]>=0.80.0
streamlit>=1.31.0
python-dotenv>=1.0.0

# Document export
python-docx>=1.1.0

# PowerPoint export (optional feature in app, but enabled if installed)
python-pptx>=0.6.23

# QR code generation (optional)
qrcode[pil]>=7.4.0